ref: rearrange emergency contact directory structure
