+++ frontend/src/features/user/emergency-contact/components/ContactNumberInput.tsx
@@ -1,18 +1,7 @@
 import { useMemo, useState } from 'react'
 import { Controller, useForm } from 'react-hook-form'
 import { BiCheck } from 'react-icons/bi'
-import {
-  Flex,
-  FormControl,
-  Modal,
-  ModalBody,
-  ModalCloseButton,
-  ModalContent,
-  ModalHeader,
-  ModalOverlay,
-  Text,
-  useDisclosure,
-} from '@chakra-ui/react'
+import { Flex, FormControl } from '@chakra-ui/react'
 
 import { isMobilePhoneNumber } from '~shared/utils/phone-num-validation'
 
@@ -31,7 +20,7 @@ type ContactNumberFormInputs = {
   contact: string
 }
 
-const ContactNumberInput = () => {
+export const ContactNumberInput = (): JSX.Element => {
   const { user, isLoading } = useUser()
   const [isVfnOpen, setIsVfnOpen] = useState(false)
   const [isSuccess, setIsSuccess] = useState(false)
@@ -132,29 +121,5 @@ const ContactNumberInput = () => {
         ) : null}
       </FormControl>
     </form>
-  )
-}
-
-export const EmergencyContactModal = (): JSX.Element => {
-  const { isOpen, onOpen, onClose } = useDisclosure()
-  return (
-    <>
-      <Button onClick={onOpen}>Open</Button>
-      <Modal isOpen={isOpen} onClose={onClose}>
-        <ModalOverlay />
-        <ModalContent>
-          <ModalCloseButton />
-          <ModalHeader color="secondary.700">Emergency contact</ModalHeader>
-          <ModalBody whiteSpace="pre-line" pb="3.25rem">
-            <Text textStyle="body-2" color="secondary.500">
-              Update your mobile number and verify it so we can contact you in
-              the unlikely case of an urgent form issue. This number can be
-              changed at any time in your user settings.
-            </Text>
-            <ContactNumberInput />
-          </ModalBody>
-        </ModalContent>
-      </Modal>
-    </>
   )
 }
